Room removed. SqlDelight working on Android.

+++ composeApp/src/commonMain/kotlin/domain/use_cases/mappers/Mappers.kt
@@ -1,8 +1,8 @@
 
 package domain.use_cases.mappers
 
-import data.database.room.tables.DbGame
-import data.database.room.tables.DbRound
+import com.etologic.mahjongscoring.DbGame
+import com.etologic.mahjongscoring.DbRound
 import domain.model.UiGame
 import domain.model.UiRound
 
@@ -22,11 +22,11 @@ fun DbGame.toUiGame(dbRounds: List<DbRound>): UiGame =
                 roundId = dbRound.roundId,
                 winnerInitialSeat = dbRound.winnerInitialSeat,
                 discarderInitialSeat = dbRound.discarderInitialSeat,
-                handPoints = dbRound.handPoints,
-                penaltyP1 = dbRound.penaltyP1,
-                penaltyP2 = dbRound.penaltyP2,
-                penaltyP3 = dbRound.penaltyP3,
-                penaltyP4 = dbRound.penaltyP4,
+                handPoints = dbRound.handPoints.toInt(),
+                penaltyP1 = dbRound.penaltyP1.toInt(),
+                penaltyP2 = dbRound.penaltyP2.toInt(),
+                penaltyP3 = dbRound.penaltyP3.toInt(),
+                penaltyP4 = dbRound.penaltyP4.toInt(),
             )
         },
     )
